requirements.txt
archspec @ file:///croot/archspec_1697725767277/work
blinker==1.7.0
boltons @ file:///C:/ci_311/boltons_1677729932371/work
Brotli @ file:///C:/ci_311/brotli-split_1676435766766/work
certifi @ file:///C:/b/abs_91u83siphd/croot/certifi_1700501720658/work/certifi
cffi @ file:///C:/b/abs_924gv1kxzj/croot/cffi_1700254355075/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama @ file:///C:/ci_311/colorama_1676422310965/work
conda @ file:///C:/b/abs_88drkd35d4/croot/conda_1701719561256/work
conda-content-trust @ file:///C:/b/abs_e3bcpyv7sw/croot/conda-content-trust_1693490654398/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1702997573971/work/src
conda-package-handling @ file:///C:/b/abs_b9wp3lr1gn/croot/conda-package-handling_1691008700066/work
conda_package_streaming @ file:///C:/b/abs_6c28n38aaj/croot/conda-package-streaming_1690988019210/work
cryptography @ file:///C:/b/abs_e8cnom_zw_/croot/cryptography_1702071486468/work
distro @ file:///C:/b/abs_a3uni_yez3/croot/distro_1701455052240/work
Flask==3.0.1
HTMLParser==0.0.2
idna @ file:///C:/ci_311/idna_1676424932545/work
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
libmambapy @ file:///C:/b/abs_efpsdwt4ya/croot/mamba-split_1698782663578/work/libmambapy
lingua-language-detector==2.0.2
MarkupSafe==2.1.4
menuinst @ file:///C:/b/abs_e8p75b4m9q/croot/menuinst_1702390332729/work
packaging @ file:///C:/b/abs_28t5mcoltc/croot/packaging_1693575224052/work
platformdirs @ file:///C:/b/abs_b6z_yqw_ii/croot/platformdirs_1692205479426/work
pluggy @ file:///C:/ci_311/pluggy_1676422178143/work
pycosat @ file:///C:/b/abs_31zywn1be3/croot/pycosat_1696537126223/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
requests @ file:///C:/b/abs_316c2inijk/croot/requests_1690400295842/work
ruamel.yaml @ file:///C:/ci_311/ruamel.yaml_1676439214109/work
tqdm @ file:///C:/b/abs_f76j9hg7pv/croot/tqdm_1679561871187/work
truststore @ file:///C:/b/abs_55z7b3r045/croot/truststore_1695245455435/work
urllib3 @ file:///C:/b/abs_9cmlsrm3ys/croot/urllib3_1698257595508/work
Werkzeug==3.0.1
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
zstandard==0.19.0
